milestone_3 update
Created a while loop which asks the user to input a single character from the alphabet when prompted. If the input doesn't match the required criteria the user is asked again to input a valid input

Then imported the random word module.

Used an if block to check if a the character the user inputed is contained within a randomly generated word using the rnadom word module